New version with FPGA as instrument
