--- src/qibolab/runcards/tii1q_rfsoc.yml
@@ -1,70 +1,0 @@
-nqubits: 1
-description: TII's 1-qubit device controlled with the RFSoC4x2 (fpga board).
-
-qubits: [0]
-
-topology: # qubit - qubit connections
--   [1]
-
-qubit_channel_map: [[0, 1]] # [readout, drive]
-# TODO: Remove repetition because these channels are defined here
-# and in the config below
-instruments:
-    tii_rfsoc4x2:
-        lib: tii
-        class: tii_rfsoc4x2
-#        address: 192.168.2.72
-        roles: [control, readout]
-        settings:
-            # Connextion
-            ip_address: 192.168.2.72
-            ip_port: 6000
-
-            # channel config
-            res_ch: 0       # Repeated at qubit_channel_map: [[0, 1]]
-            qubit_ch: 1
-            # resonator config
-            readout_length: 3.0 # Micro seg.
-            f_res: 7295.23829 # [MHz]
-            resonator_phase: 3097210280 #
-            adc_trig_offset: 200 # [Clock ticks]
-            res_gain: 600 #
-            threshold: 0
-            # qubit config
-            sigma: 0.114 # Micro seg.
-            pi_gain: 17000 #
-            pi2_gain: 17000 // 2
-            f_ge: 5236.3 # [MHz]
-            relax_delay: 500 #[Clock ticks]
-        experiment:
-            reps: 1024
-            max_gain: 30000
-            delay_before_readout: 50 # [ns]
-            rabi_length: 0.01
-native_gates:
-    single_qubit:
-        0: # qubit number
-            RX:
-                duration: 114
-                amplitude: 0.566
-                frequency: 5_236_300_000
-                shape: Drag(5, 0.05096)
-                type: qd # qubit drive
-            MZ:
-                duration: 3000
-                amplitude: 0.02
-                frequency: 7_295_238_290
-                shape: Rectangular()
-                type: ro # readout
-
-characterization:
-    single_qubit:
-        0:
-            resonator_freq: 7824847574
-            qubit_freq: 5082293989
-            T1: 17744
-            T2: 8385
-            state0_voltage: 891
-            state1_voltage: 357
-            mean_gnd_states: (-5.543255134868772e-05+0.0008975930158821524j)
-            mean_exc_states: (0.0001446400604158933+0.0003369969077198778j)
