+++ src/qibolab/runcards/tii_rfsoc4x2.yml
@@ -13,7 +13,7 @@ instruments:
     tii_rfsoc4x2:
         lib: tii
         class: tii_rfsoc4x2
-#        address: 192.168.2.72
+        address: 192.168.2.72
         roles: [control, readout]
         settings:
             # Connextion
@@ -24,14 +24,14 @@ instruments:
             res_ch: 0       # Repeated at qubit_channel_map: [[0, 1]]
             qubit_ch: 1
             # resonator config
-            readout_length: 3.0 # Micro seg.
-            f_res: 7295.23829 # [MHz]
+            readout_length: 3000 # ns
+            f_res: 7_295_238_290 # [Hz]
             resonator_phase: 3097210280 #
             adc_trig_offset: 200 # [Clock ticks]
             res_gain: 600 #
             threshold: 0
             # qubit config
-            sigma: 0.114 # Micro seg.
+            sigma: 114 # ns
             pi_gain: 17000 #
             pi2_gain: 17000 // 2
             f_ge: 5236.3 # [MHz]
@@ -56,3 +56,15 @@ native_gates:
                 frequency: 7_295_238_290
                 shape: Rectangular()
                 type: ro # readout
+
+characterization:
+    single_qubit:
+        0:
+            resonator_freq: 7824847574
+            qubit_freq: 5082293989
+            T1: 17744
+            T2: 8385
+            state0_voltage: 891
+            state1_voltage: 357
+            mean_gnd_states: (-5.543255134868772e-05+0.0008975930158821524j)
+            mean_exc_states: (0.0001446400604158933+0.0003369969077198778j)
